+++ packages/client/base/src/services/api/APIErrorService.ts
@@ -1,5 +1,6 @@
+import { CrossmintSDKError } from "@crossmint/common-sdk-base";
+
 import {
-    CrossmintSDKError,
     CrossmintServiceError,
     JWTDecryptionError,
     JWTExpiredError,
