+++ packages/client/base/src/types/index.ts
@@ -1,4 +1,3 @@
-export * from "./errors";
 export * from "./events";
 export * from "./hosted";
 export * from "./embed";
