+++ packages/common/base/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./apiClient";
 export * from "./apiKey";
 export * from "./blockchain";
+export * from "./error";
 export * from "./types";
