+++ packages/common/base/src/types/errors.ts
