+++ packages/common/base/src/types/index.ts
@@ -1,4 +1,5 @@
 export * from "./Crossmint";
+export * from "./errors";
 export * from "./react";
 export * from "./uiconfig";
 export * from "./UserIdentifier";
